Fix component import and update language parameter in SavedMeals component
