+++ src/app/[lng]/components/storyBookComponents/history/historyCard.tsx
@@ -1,4 +1,4 @@
-'use client';
+'use client'
 import React from 'react';
 import './historyCard.css';
 import '/src/app/globals.css';
